requirements.txt
Django==1.8.2
django-bootstrap-form==3.2
django-cors-headers==1.1.0
django-datatable-view==0.8.2
django-easy-pdf==0.1.0
django-helpdesk==0.1.16
django-markdown-deux==1.0.5
django-templated-email==0.4.9
email-reply-parser==0.3.0
html5lib==0.999999
markdown2==2.3.0
MySQL-python==1.2.5
Pillow==2.9.0
PyPDF2==1.25.1
python-dateutil==2.4.2
pytz==2015.4
reportlab==2.7
simplejson==3.8.0
six==1.9.0
wheel==0.24.0
xhtml2pdf==0.0.6
django-tinymce==2.0.6
gunicorn==19.4.1
redis==2.10.5
celery==3.1.18
django-phonenumber-field==1.0.0
ujson==1.34
django-bootstrap3==7.0.1
